Move `render_orthographic()` code to render crate.
It's still very much a special case, but we can now start on
more integration or using it elsewhere than the UI test.
